fix(runtime,cli): surface nested RPC errors, drain readers fully, preflight the worker, split worker-op tests
- WorkerClient.request reads error.data.message (RuntimeServer's channel
  for RuntimeApiError's actionable text) with a fallback to the top-level
  message, so  failures name the failed step instead of a bare
  command_failed.
- archive_own_step_logs teardown joins the reader a second time after the
  stop signal and records a TimeoutError on the reader state when the
  drain never completes, instead of releasing the guard over a live
  reader.
-  preflights worker availability before creating the
  run directory (and before --overwrite deletes an existing one), so a
  missing worker leaves the workspace untouched for a clean retry; the
  orphaned duplicate of the check in project.py is gone.
- test_worker_operation.py splits into lifecycle/sequencing, crash
  recovery, and archive modules over a shared worker_operation_support
  preamble, each below the 700-line review bar.
Regressions pin the nested-message parsing, the preflight (run dir
uncreated; overwrite target preserved), and all split suites pass.
