+++ test/runtime/test_worker_operation.py
@@ -1,67 +1,14 @@
-"""Tests for chipcompiler.runtime.worker_operation — typed operation orchestrator."""
+"""Tests for RunOperation session lifecycle and sequencing."""
 
-import json
 import os
 import sys
 import textwrap
 
 import pytest
+from worker_operation_support import _RPC_HELPERS, LIFECYCLE_SERVER
 
 from chipcompiler.runtime.worker import WorkerClient
 from chipcompiler.runtime.worker_operation import RunOperation
-
-# Common RPC helpers injected into subprocess scripts.
-_RPC_HELPERS = textwrap.dedent("""\
-    import sys, os, json
-
-    def read_request():
-        data = b""
-        while True:
-            chunk = sys.stdin.buffer.read(1)
-            if not chunk:
-                return None
-            data += chunk
-            if b"\\r\\n\\r\\n" in data:
-                header, _, body_start = data.partition(b"\\r\\n\\r\\n")
-                length = int(header.split(b":")[1])
-                while len(body_start) < length:
-                    body_start += sys.stdin.buffer.read(1)
-                return json.loads(body_start[:length])
-
-    def send_response(resp):
-        payload = json.dumps(resp)
-        frame = f"Content-Length: {len(payload)}\\r\\n\\r\\n{payload}"
-        sys.stdout.buffer.write(frame.encode())
-        sys.stdout.buffer.flush()
-
-    def make_marker(event, step, tool):
-        p = json.dumps({"v": 1, "event": event, "step": step, "tool": tool})
-        return chr(0x1e).encode() + b"ECC-STEP " + p.encode() + b"\\n"
-""")
-
-LIFECYCLE_SERVER = _RPC_HELPERS + textwrap.dedent("""\
-    while True:
-        req = read_request()
-        if req is None:
-            break
-        method = req.get("method", "")
-        req_id = req.get("id")
-
-        if method == "rpc.hello":
-            resp = {"jsonrpc": "2.0", "result": {"version": 1, "capabilities": []}, "id": req_id}
-            send_response(resp)
-        elif method == "workspace.open":
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "test"}, "id": req_id})
-        elif method == "rpc.shutdown":
-            send_response({"jsonrpc": "2.0", "result": {"ok": True}, "id": req_id})
-            break
-        elif method == "flow.run":
-            send_response({"jsonrpc": "2.0", "result": {"steps": ["syn"]}, "id": req_id})
-        else:
-            err = {"code": -32601, "message": "unknown method"}
-            resp = {"jsonrpc": "2.0", "error": err, "id": req_id}
-            send_response(resp)
-""")
 
 
 class TestRunOperationLifecycle:
@@ -183,405 +130,6 @@ class TestRunOperationSequence:
         received = log_file.read_text().splitlines()
         assert "flow.run" not in received
         assert received[-1] == "rpc.shutdown"
-
-
-class TestRunOperationRpcErrorRepair:
-    def test_rpc_error_with_unmatched_step_repairs_flow_state(self, tmp_path):
-        """A live worker's RPC error still repairs a step left Ongoing."""
-        script = tmp_path / "error_after_begin.py"
-        script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            while True:
-                req = read_request()
-                if req is None:
-                    break
-                method = req.get("method", "")
-                req_id = req.get("id")
-                if method == "rpc.hello":
-                    send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req_id})
-                elif method == "workspace.open":
-                    result = {"workspaceId": "x"}
-                    send_response({"jsonrpc": "2.0", "result": result, "id": req_id})
-                elif method == "flow.run_step":
-                    os.write(2, make_marker("begin", "Synthesis", "yosys"))
-                    os.write(2, b"partial output\\n")
-                    err = {"code": -32000, "message": "run step Synthesis failed"}
-                    send_response({"jsonrpc": "2.0", "error": err, "id": req_id})
-                elif method == "rpc.shutdown":
-                    send_response({"jsonrpc": "2.0", "result": {"ok": True}, "id": req_id})
-                    break
-                else:
-                    err = {"code": -32601, "message": "unknown method"}
-                    send_response({"jsonrpc": "2.0", "error": err, "id": req_id})
-        """)
-        )
-        flow_json = tmp_path / "flow.json"
-        data = {"steps": [{"name": "Synthesis", "tool": "yosys", "state": "Ongoing"}]}
-        flow_json.write_text(json.dumps(data))
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(script)],
-        )
-        result = op.run_sequence([("flow.run_step", {"step": "Synthesis", "rerun": True})])
-        assert result.success is False
-        assert "run step Synthesis failed" in result.error
-        assert result.repaired_steps == ["Synthesis"]
-        repaired_data = json.loads(flow_json.read_text())
-        assert repaired_data["steps"][0]["state"] == "Incomplete"
-
-
-class TestRunOperationCrash:
-    def test_worker_crash_triggers_repair(self, tmp_path):
-        crash_script = tmp_path / "crash_after_open.py"
-        crash_script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "x"}, "id": req["id"]})
-            # Emit begin marker on stderr, then crash
-            os.write(2, make_marker("begin", "Synthesis", "yosys"))
-            os._exit(1)
-        """)
-        )
-        flow_json = tmp_path / "flow.json"
-        data = {"steps": [{"name": "Synthesis", "tool": "yosys", "state": "Ongoing"}]}
-        flow_json.write_text(json.dumps(data))
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(crash_script)],
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is False
-        assert result.repaired_steps == ["Synthesis"]
-        repaired_data = json.loads(flow_json.read_text())
-        assert repaired_data["steps"][0]["state"] == "Incomplete"
-
-    def test_crash_repair_failure_is_reported(self, tmp_path, monkeypatch):
-        """A repair that cannot persist must surface in the result error."""
-        crash_script = tmp_path / "crash_after_open.py"
-        crash_script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "x"}, "id": req["id"]})
-            os.write(2, make_marker("begin", "Synthesis", "yosys"))
-            os._exit(1)
-        """)
-        )
-        flow_json = tmp_path / "flow.json"
-        data = {"steps": [{"name": "Synthesis", "tool": "yosys", "state": "Ongoing"}]}
-        flow_json.write_text(json.dumps(data))
-
-        def failing_repair(*args, **kwargs):
-            raise OSError("disk full")
-
-        monkeypatch.setattr(
-            "chipcompiler.runtime.worker_operation.repair_flow_state", failing_repair
-        )
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(crash_script)],
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is False
-        assert result.repaired_steps == []
-        assert "state repair failed" in result.error
-        # The record is left as it was — the failure is reported, not hidden.
-        assert json.loads(flow_json.read_text())["steps"][0]["state"] == "Ongoing"
-
-    def test_parent_sigterm_terminates_the_worker_group(self, tmp_path):
-        """SIGTERM to the CLI must route through crash recovery and reap the
-        worker process group instead of leaving EDA descendants running."""
-        import signal as signal_module
-        import threading
-        import time
-
-        script = _RPC_HELPERS + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "x"}, "id": req["id"]})
-            # Hang forever, simulating a long EDA step.
-            while True:
-                time.sleep(1)
-        """)
-        script = "import time\n" + script
-        flow_json = tmp_path / "flow.json"
-        flow_json.write_text(json.dumps({"steps": []}))
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, "-c", script],
-        )
-
-        def send_sigterm():
-            time.sleep(0.5)
-            os.kill(os.getpid(), signal_module.SIGTERM)
-
-        killer = threading.Thread(target=send_sigterm, daemon=True)
-        killer.start()
-        result = op.run("flow.run", {"workspace_id": "test"})
-
-        assert result.success is False
-        assert "interrupted" in (result.error or "")
-        assert result.signal_number == -signal_module.SIGKILL or result.exit_code is not None
-
-    def test_crash_before_any_marker_repairs_persisted_ongoing(self, tmp_path):
-        """Killed between the Ongoing save and the begin marker: no stream
-        evidence exists, so recovery falls back to the persisted Ongoing."""
-        crash_script = tmp_path / "crash_before_marker.py"
-        crash_script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "x"}, "id": req["id"]})
-            # Die before any marker — as if killed between the Ongoing save
-            # and the begin write.
-            os._exit(1)
-        """)
-        )
-        flow_json = tmp_path / "flow.json"
-        flow_json.write_text(
-            json.dumps({"steps": [{"name": "Synthesis", "tool": "yosys", "state": "Ongoing"}]})
-        )
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(crash_script)],
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is False
-        assert result.repaired_steps == ["Synthesis"]
-        repaired = json.loads(flow_json.read_text())
-        assert repaired["steps"][0]["state"] == "Incomplete"
-
-    def test_worker_crash_no_flow_json_no_repair(self, tmp_path):
-        script = _RPC_HELPERS + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "x"}, "id": req["id"]})
-            sys.exit(1)
-        """)
-        flow_json = tmp_path / "flow.json"
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, "-c", script],
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is False
-        assert result.repaired_steps == []
-
-    def test_protocol_failure_routes_to_recovery(self, tmp_path):
-        """A worker that returns invalid JSON triggers crash recovery."""
-        script = _RPC_HELPERS + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "x"}, "id": req["id"]})
-            # Emit begin marker, then send garbage on stdout
-            os.write(2, make_marker("begin", "Place", "ecc"))
-            req = read_request()  # flow.run
-            # Send invalid frame (bad Content-Length header)
-            sys.stdout.buffer.write(b"Content-Length: 5\\r\\n\\r\\n{}")
-            sys.stdout.buffer.flush()
-            sys.stdout.buffer.close()
-            os._exit(1)
-        """)
-        flow_json = tmp_path / "flow.json"
-        data = {"steps": [{"name": "Place", "tool": "ecc", "state": "Ongoing"}]}
-        flow_json.write_text(json.dumps(data))
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, "-c", script],
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is False
-        assert result.repaired_steps == ["Place"]
-
-
-class TestRunOperationArchive:
-    def test_archive_error_forces_failure(self, tmp_path):
-        """Success is False when archive path cannot be opened."""
-        script = tmp_path / "server_with_markers.py"
-        script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "ws1"}, "id": req["id"]})
-            os.write(2, make_marker("begin", "Synthesis", "yosys"))
-            os.write(2, b'Synthesizing...\\n')
-            os.write(2, make_marker("end", "Synthesis", "yosys"))
-            req = read_request()  # flow.run
-            send_response({"jsonrpc": "2.0", "result": {"steps": ["syn"]}, "id": req["id"]})
-            req = read_request()  # rpc.shutdown
-            send_response({"jsonrpc": "2.0", "result": {"ok": True}, "id": req["id"]})
-        """)
-        )
-
-        # A regular file as path component makes mkdir fail with ENOTDIR even
-        # when tests run as root (chmod-based read-only setups are bypassed).
-        blocker = tmp_path / "not_a_dir"
-        blocker.write_text("regular file")
-
-        def bad_resolver(step: str, tool: str):
-            return blocker / "sub" / f"{step}.log"
-
-        flow_json = tmp_path / "flow.json"
-        flow_json.write_text("{}")
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(script)],
-            log_path_resolver=bad_resolver,
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is False
-        assert result.archive_error is not None
-        assert "archive error" in result.error
-
-    def test_display_callback_failure_is_not_an_archive_failure(self, tmp_path):
-        """A broken on_output renderer must not downgrade the step or fail
-        the operation when the archive itself is fine."""
-        script = tmp_path / "server_with_markers.py"
-        script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "ws1"}, "id": req["id"]})
-            os.write(2, make_marker("begin", "Synthesis", "yosys"))
-            os.write(2, b'Synthesizing...\\n')
-            os.write(2, make_marker("end", "Synthesis", "yosys"))
-            req = read_request()  # flow.run
-            send_response({"jsonrpc": "2.0", "result": {"steps": ["syn"]}, "id": req["id"]})
-            req = read_request()  # rpc.shutdown
-            send_response({"jsonrpc": "2.0", "result": {"ok": True}, "id": req["id"]})
-        """)
-        )
-        logs_dir = tmp_path / "logs"
-
-        def resolver(step: str, tool: str):
-            return logs_dir / f"{step}.log"
-
-        def broken_display(data: bytes):
-            raise RuntimeError("renderer blew up")
-
-        flow_json = tmp_path / "flow.json"
-        flow_json.write_text(
-            json.dumps({"steps": [{"name": "Synthesis", "tool": "yosys", "state": "Success"}]})
-        )
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(script)],
-            log_path_resolver=resolver,
-            on_output=broken_display,
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is True
-        assert result.repaired_steps == []
-        assert result.log_state is not None
-        assert result.log_state.display_error is not None
-        assert result.log_state.error is None
-        assert json.loads(flow_json.read_text())["steps"][0]["state"] == "Success"
-
-    def test_archive_error_reconciles_the_success_record(self, tmp_path):
-        """An archive failure must not leave flow.json claiming Success."""
-        script = tmp_path / "server_with_markers.py"
-        script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "ws1"}, "id": req["id"]})
-            os.write(2, make_marker("begin", "Synthesis", "yosys"))
-            os.write(2, b'Synthesizing...\\n')
-            os.write(2, make_marker("end", "Synthesis", "yosys"))
-            req = read_request()  # flow.run
-            send_response({"jsonrpc": "2.0", "result": {"steps": ["syn"]}, "id": req["id"]})
-            req = read_request()  # rpc.shutdown
-            send_response({"jsonrpc": "2.0", "result": {"ok": True}, "id": req["id"]})
-        """)
-        )
-
-        blocker = tmp_path / "not_a_dir"
-        blocker.write_text("regular file")
-
-        def bad_resolver(step: str, tool: str):
-            return blocker / "sub" / f"{step}.log"
-
-        flow_json = tmp_path / "flow.json"
-        flow_json.write_text(
-            json.dumps({"steps": [{"name": "Synthesis", "tool": "yosys", "state": "Success"}]})
-        )
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(script)],
-            log_path_resolver=bad_resolver,
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is False
-        assert result.archive_error is not None
-        # The persisted record is downgraded so a later resume reruns the step
-        # and recreates the missing log instead of skipping it.
-        assert result.repaired_steps == ["Synthesis"]
-        repaired = json.loads(flow_json.read_text())
-        assert repaired["steps"][0]["state"] == "Incomplete"
-
-    def test_stderr_archived_to_step_log(self, tmp_path):
-        script = tmp_path / "server_with_markers.py"
-        script.write_text(
-            _RPC_HELPERS
-            + textwrap.dedent("""\
-            req = read_request()  # hello
-            send_response({"jsonrpc": "2.0", "result": {"version": 1}, "id": req["id"]})
-            req = read_request()  # workspace.open
-            send_response({"jsonrpc": "2.0", "result": {"workspaceId": "ws1"}, "id": req["id"]})
-            os.write(2, make_marker("begin", "Synthesis", "yosys"))
-            os.write(2, b'Synthesizing module top...\\n')
-            os.write(2, make_marker("end", "Synthesis", "yosys"))
-            req = read_request()  # flow.run
-            send_response({"jsonrpc": "2.0", "result": {}, "id": req["id"]})
-            req = read_request()  # rpc.shutdown
-            send_response({"jsonrpc": "2.0", "result": {"ok": True}, "id": req["id"]})
-        """)
-        )
-        logs_dir = tmp_path / "logs"
-
-        def resolver(step: str, tool: str):
-            return logs_dir / f"{step}.log"
-
-        flow_json = tmp_path / "flow.json"
-        flow_json.write_text("{}")
-        op = RunOperation(
-            workspace_dir=tmp_path,
-            flow_json_path=flow_json,
-            worker_argv=[sys.executable, str(script)],
-            log_path_resolver=resolver,
-        )
-        result = op.run("flow.run", {"workspace_id": "test"})
-        assert result.success is True
-        log_file = logs_dir / "Synthesis.log"
-        assert log_file.exists()
-        assert b"Synthesizing module top..." in log_file.read_bytes()
 
 
 class TestWorkspaceIdInjection:
